% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{color_blue}
\alias{color_blue}
\title{Column coloring}
\usage{
color_blue(col_index, data, ft)
}
\arguments{
\item{col_index}{The index of the column to color}

\item{data}{The dataframe that created the flextable.}

\item{ft}{A flextable.}
}
\description{
This function colors a column of a flextable blue.
}
