% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggpairs.R
\name{plot_ggpairs}
\alias{plot_ggpairs}
\title{Helper function for GGally::ggpairs}
\usage{
plot_ggpairs(data, color = NULL, density_alpha = 0.5)
}
\arguments{
\item{data}{data frame containing features and target}

\item{color}{column name to use for color aesthetic in pair plots}

\item{density_alpha}{alpha setting for frequency density plots}
}
\value{
ggplot object containing pairwise comparisons
}
\description{
Helper function for GGally::ggpairs
}
\examples{
plot_ggpairs(iris)
}
