% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_facet.R
\name{plot_hist_facet}
\alias{plot_hist_facet}
\title{Faceted ggplot histograms}
\usage{
plot_hist_facet(data, bins = 10, ncol = 5, fct_reorder = FALSE,
  fct_rev = FALSE, fill = "#18BC9C", color = "white",
  scale = "free")
}
\arguments{
\item{data}{data frame containing features}

\item{bins}{number of histogram bins}

\item{ncol}{number of histogram columns}

\item{fct_reorder}{reorder factors}

\item{fct_rev}{reverse factor order}

\item{fill}{color value for geom_hist fill argument}

\item{color}{color value for geom_hist color argument}

\item{scale}{value for geom_hist scale argument

Original code by Matt Dancho DS4B 201}
}
\value{
ggplot object containing histograms
}
\description{
Faceted ggplot histograms
}
\examples{
plot_hist_facet(iris)
}
