% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\docType{package}
\name{viz-package}
\alias{viz}
\alias{viz-package}
\title{viz: visualization helper functions, themes, and palettes}
\description{
Helper functions:
}
\details{
\itemize{
\item plot_ggpairs: GGally::ggpairs helper
}

Themes:

Palettes:
}
